refactoring createHiddenInput
