+++ app/javascript/controllers/projects/samples/attachments/selected_attachments_controller.js
@@ -1,4 +1,5 @@
 import { Controller } from "@hotwired/stimulus";
+import { createHiddenInput } from '../../../../utilities/form';
 
 //creates hidden fields within a form for selected files
 export default class extends Controller {
@@ -28,15 +29,13 @@ export default class extends Controller {
 
         if (value instanceof Array) {
           for (let arrayValue of value) {
-            this.#addHiddenInput(
-              `${this.fieldNameValue}[${storageValueIndex}][]`,
-              arrayValue
+            this.fieldTarget.appendChild(
+              createHiddenInput(`${this.fieldNameValue}[${storageValueIndex}][]`, arrayValue)
             );
           }
         } else {
-          this.#addHiddenInput(
-            `${this.fieldNameValue}[${storageValueIndex}]`,
-            value
+          this.fieldTarget.appendChild(
+            createHiddenInput(`${this.fieldNameValue}[${storageValueIndex}]`, value)
           );
         }
       }
@@ -47,15 +46,5 @@ export default class extends Controller {
     if (event.detail.success) {
       sessionStorage.removeItem(this.storageKeyValue);
     }
-  }
-
-  #addHiddenInput(name, value) {
-    const element = document.createElement("input");
-    element.type = "hidden";
-    element.id = value;
-    element.name = name;
-    element.value = value;
-    element.ariaHidden = "true";
-    this.fieldTarget.appendChild(element);
   }
 }
