+++ app/javascript/controllers/sessionstorage_amend_form_controller.js
@@ -1,4 +1,5 @@
 import { Controller } from "@hotwired/stimulus";
+import { createHiddenInput } from '../utilities/form';
 
 export default class extends Controller {
   static targets = ["field"];
@@ -19,7 +20,7 @@ export default class extends Controller {
     if (storageValues) {
       const fragment = document.createDocumentFragment();
       storageValues.forEach((value) => {
-        fragment.appendChild(this.#createHiddenInput(value));
+        fragment.appendChild(createHiddenInput(this.fieldNameValue, value));
       });
       this.fieldTarget.appendChild(fragment);
     }
@@ -27,13 +28,5 @@ export default class extends Controller {
 
   clear() {
     sessionStorage.removeItem(this.storageKeyValue);
-  }
-
-  #createHiddenInput(value) {
-    const element = document.createElement("input");
-    element.type = "hidden";
-    element.name = this.fieldNameValue;
-    element.value = value;
-    return element;
   }
 }
